refactor: 💡 chart utils
