+++ packages/charts/src/charts/bubble/index.ts
@@ -1,6 +1,7 @@
 import BubbleChart, {
   Props as BubbleChartSettings,
 } from './bubble-chart.component';
+import { bubbleColorScale } from './utils';
 
-export { BubbleChart };
+export { BubbleChart, bubbleColorScale };
 export type { BubbleChartSettings };
