+++ packages/charts/src/charts/bubble/utils/index.ts
@@ -1,0 +1,4 @@
+import { randomizeVector } from './animation.utils';
+import { generateBubbles, bubbleColorScale } from './chart.utils';
+
+export { randomizeVector, generateBubbles, bubbleColorScale };
