+++ packages/charts/src/charts/circular-chart/constants.ts
@@ -1,0 +1,1 @@
+export const OTHERS_DATA_KEY = 'Others';
