+++ packages/charts/src/charts/circular-chart/index.ts
@@ -1,3 +1,23 @@
 import TooltipItem from './tooltip-item.component';
+import {
+  generateCircularChart,
+  getCircularChartTooltipContent,
+  createArcTween,
+  animateArcPath,
+} from './utils';
+import { OTHERS_DATA_KEY } from './constants';
+import {
+  LabelsPosition,
+  Options as CircularChart,
+  ArcProperties,
+} from './types';
 
-export { TooltipItem };
+export type { LabelsPosition, CircularChart, ArcProperties };
+export {
+  TooltipItem,
+  generateCircularChart,
+  getCircularChartTooltipContent,
+  createArcTween,
+  animateArcPath,
+  OTHERS_DATA_KEY,
+};
