+++ packages/charts/src/charts/circular-chart/utils/chart.utils.ts
@@ -1,67 +1,14 @@
 import { sum } from 'd3-array';
 import { arc, pie } from 'd3-shape';
-import { colors as palette } from '@keen.io/colors';
 import {
   calculateHypotenuse,
-  Formatter,
   formatValue as valueFormatter,
   getOffsetRangeColor,
 } from '@keen.io/charts-utils';
+import { colors as palette } from '@keen.io/colors';
 
-import { Dimension, Margins, DataSelector } from '../types';
-
-export type LabelsPosition = 'inside' | 'outside';
-
-type SliceType = 'pie' | 'donut';
-
-export type Options = {
-  data: Record<string, any>[];
-  labelSelector: string;
-  keys: string[];
-  disabledLabels: string[];
-  dimension: Dimension;
-  margins: Margins;
-  colors: string[];
-  padAngle: number;
-  padRadius: number;
-  cornerRadius: number;
-  innerRadius: number;
-  labelsRadius: number;
-  labelsPosition: LabelsPosition;
-  type?: SliceType;
-  treshold?: number;
-  formatValue?: Formatter;
-  dataSeriesOffset?: [number, number];
-};
-
-type Arc = {
-  index: string;
-  labelNumeric: string | number | Date | boolean;
-  labelPercentage: string;
-  activePosition: [number, number];
-  labelPosition: [number, number];
-  color: string;
-  selector: DataSelector;
-  dataKey: string;
-  startAngle: number;
-  endAngle: number;
-  stacked: boolean;
-  stack: { selector: DataSelector; color: string }[];
-};
-
-type Slice = {
-  color: string;
-  value: number;
-  dataKey: string;
-  selector: DataSelector;
-  stacked?: boolean;
-  stack?: { selector: DataSelector; color: string }[];
-  index?: number;
-};
-
-export const HOVER_RADIUS = 5;
-
-export const OTHERS_DATA_KEY = 'Others';
+import { OTHERS_DATA_KEY } from '../constants';
+import { Arc, Options, Slice } from '../types';
 
 export const createStackedSlice = ({
   slices,
