+++ packages/charts/src/charts/metric/utils/element.utils.ts
@@ -1,4 +1,4 @@
-import { Margins } from '../types';
+import { Margins } from '../../../types';
 
 export const createMargins = ({ top, bottom, left, right }: Margins) => ({
   marginTop: `${top}px`,
