+++ packages/charts/src/charts/metric/utils/index.ts
@@ -1,5 +1,6 @@
 import { generateMetric } from './metric.utils';
 import { formatNumber } from './format.utils';
 import { generateCircles } from './icon.utils';
+import { createMargins } from './element.utils';
 
-export { generateMetric, generateCircles, formatNumber };
+export { generateMetric, generateCircles, formatNumber, createMargins };
