+++ packages/charts/src/utils/index.ts
@@ -1,23 +1,4 @@
-import { createArcTween, animateArcPath, ArcProperties } from './animate.utils';
-import { bubbleColorScale } from './scale.utils';
 import { calculateColorScale } from './colors.utils';
 import { extendTheme } from './extend-theme';
 
-import {
-  generateCircularChart,
-  LabelsPosition,
-  Options as CircularChart,
-  OTHERS_DATA_KEY,
-} from './circular-chart.utils';
-
-export type { ArcProperties, CircularChart, LabelsPosition };
-
-export {
-  createArcTween,
-  animateArcPath,
-  generateCircularChart,
-  OTHERS_DATA_KEY,
-  bubbleColorScale,
-  calculateColorScale,
-  extendTheme,
-};
+export { calculateColorScale, extendTheme };
