--- packages/charts/src/utils/scale.utils.ts
@@ -1,3 +1,0 @@
-import { scaleOrdinal } from 'd3-scale';
-
-export const bubbleColorScale = (colors: string[]) => scaleOrdinal(colors);
