--- packages/charts/src/utils/text/first-capital.test.ts
@@ -1,6 +1,0 @@
-import { firstCapital } from './first-capital';
-
-test('capitalizes first letter in word', () => {
-  const result = firstCapital('category');
-  expect(result).toEqual('Category');
-});
