--- packages/charts/src/utils/text/first-capital.ts
@@ -1,2 +1,0 @@
-export const firstCapital = (str: string) =>
-  str.replace(/^\w/, (c: string) => c.toUpperCase());
