--- packages/charts/src/utils/text/index.ts
@@ -1,3 +1,0 @@
-import { firstCapital } from './first-capital';
-
-export { firstCapital };
